Fix scene applying for devices with assumed state
The scene applying got broken, recently. Apparently the code got
optimized to not trigger state updates if a current state is matching
the target state. Sadly, there is a race condition in that behavior,
because a device state in HA is usually just a cache and it is not
authoritative. The problem is super visible with devices that have
assumed state, but it exist in all types of entities excluding ones
fully owned by HA, like for example; automations, groups,
input_numbers etc.

This patch fixes _only_ devices that have assumed state attribute,
every other device will be affected by the race condition mentioned
earlier.